requirements.txt
python-decouple
discord
selenium
webdriver_manager